Added Individual_Levels.json
Added a config file that supports running individual levels.
To use, set the game name to "" (Blank, nothing), and use category to
set the run title for saving.
Set the Autosplitter split config to the level exit type for the level.
The autostart is set to start when selecting a level from the overworld.